support for human friendly enum labels during UpdateStatus (serialization)
when a CR is created for the 1st time, it will unless there is custom serializor, System.Text.json won't be able to save it to the custom enum label. Using the custom enum label, it can now save the enum as their custom label.

Issue - Once the resource is saved, the next time the watcher reads this resource it wont know how to deserialize the custom enum, so next we need to also add the custom enum serialization/deserialization logic for watcher